+++ rootcmd/root.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/hazelcast/hazelcast-commandline-client/clustercmd"
 	"github.com/hazelcast/hazelcast-commandline-client/config"
-	"github.com/hazelcast/hazelcast-commandline-client/listcmd"
+	"github.com/hazelcast/hazelcast-commandline-client/list_objects"
 	"github.com/hazelcast/hazelcast-commandline-client/sqlcmd"
 	fakeDoor "github.com/hazelcast/hazelcast-commandline-client/types/fakedoorcmd"
 	"github.com/hazelcast/hazelcast-commandline-client/types/mapcmd"
@@ -73,7 +73,7 @@ func subCommands(config *hazelcast.Config) []*cobra.Command {
 		mapcmd.New(config),
 		sqlcmd.New(config),
 		versioncmd.New(),
-		listcmd.New(config),
+		list_objects.New(config),
 	}
 	fds := []fakeDoor.FakeDoor{
 		{Name: "List", IssueNum: 48},
